+++ docs/source/changelog.rst
@@ -15,6 +15,7 @@ Version (dev)
 * Add :py:class:`srctools.Keyvalues.LeafKeyvalueError`, raised when block-only operations are
   attempted on leaf keyvalues. This improves the messages raised and makes them consistent.
 * Fix :py:class:`srctools.vtf.Frame` indexing behaviour. It would access totally incorrect pixels.
+* Correctly read/write L4D2's BSP format.
 
 -------------
 Version 2.3.10